removed MUST_ASSERT macro, details in description
TL;DR the new ASSERT is like MUST_ASSERT but better.

basically I originally created MUST_ASSERT as an assert which will stay in runtime, so you can keep the checks in the final application as an additional layer of protection. I originally wasn't sure if this was the best course of action, so I discussed this with many friends and I've analyzed alternatives.
The new changes to ASSERT macros (aka the returned expression values when the macro is deactivated) now seem to me a better alternative, because you have the choice of keeping the runtime check but on a case per case basis, instead of a "you get all or nothing" alternative that MUST_ASSERT provided. Moreover the intentional abort is no more the only alternative, you can use the new ASSERT and the application will continue running instead of crashing.